\name{addalpha}
\alias{addalpha}

\title{
 add alpha for color
}

\description{
 add alpha for color
}

\usage{
  addalpha(colors, alpha=1.0);
}

\arguments{
   \item{colors}{colors}
   \item{alpha}{alpha}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

