\name{divide_large_graph}
\alias{divide_large_graph}

\title{
 divide large graph
}

\description{
 divide large graph
}

\usage{
  divide_large_graph(graph, max_nodes=10);
}

\arguments{
   \item{graph}{igraph}
   \item{max_nodes}{the maximum number of nodes in graph}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

