\name{g2db}
\alias{g2db}

\title{
 convert graph list to database for GSEA
}

\description{
 convert graph list to database for GSEA
}

\usage{
  g2db(g_list, creator=creator, version=version, mydate=NULL);
}

\arguments{
   \item{g_list}{list: igraph object list}
   \item{creator}{user name}
   \item{version}{the database version}
   \item{mydate}{data}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

