\name{g2subg}
\alias{g2subg}

\title{
 divide large graph
}

\description{
 divide large graph
}

\usage{
  g2subg(g, max_nodes=20, compon=FALSE);
}

\arguments{
   \item{g}{igraph object}
   \item{max_nodes}{maximum number of nodes in network}
   \item{compon}{logic, do components only if TRUE}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

