\name{l2db}
\alias{l2db}

\title{
 convert gene list to database for GSEA
}

\description{
 convert gene list to database for GSEA
}

\usage{
  l2db(gene_list, creator=creator, version=version, mydate=NULL);
}

\arguments{
   \item{gene_list}{list: gene sets}
   \item{creator}{user name}
   \item{version}{the database version}
   \item{mydate}{data}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

