\name{l2g}
\alias{l2g}

\title{
 convert gene list to igraph
}

\description{
 convert gene list to igraph
}

\usage{
  l2g(gene, relation, INTXN=TRUE);
}

\arguments{
   \item{gene}{character: gene vetor}
   \item{relation}{gene relationship, such as PPI}
   \item{INTXN}{logic: if TRUE, do intersct}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

