\name{ppi2net}
\alias{ppi2net}

\title{
 convert ppi to network
}

\description{
 convert ppi to network
}

\usage{
  ppi2net(data);
}

\arguments{
   \item{data}{Columns, 1 and 2 are a pair of interacting genes.}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

