%% main
\name{SubPath-package}
\alias{SubPath}
\title{
 SubPath package
}
\description{
 SubPath tools.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}
\references{
%%  ~~ possibly secondary sources and usages ~~
 SubPath: an R package to divide large graph
}
%%\keyword{key}
