\name{textplot2}
\alias{textplot2}

\title{
 text plot
}

\description{
 text plot
}

\usage{
  textplot2(x, y, words, cex = 1, new = TRUE, show.lines = TRUE, ...);
}

\arguments{
   \item{x}{x}
   \item{y}{y}
   \item{words}{text}
   \item{cex}{cex}
   \item{new}{new}
   \item{show.lines}{show.lines}
   \item{...}{...}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

